backend/requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
secure-smtplib
streamlit
requests
